+++ tests/ccsds/test_time.py
@@ -68,3 +68,17 @@ class TestTime(TestCase):
         stamp += datetime.timedelta(milliseconds=10)
         self.assertEqual(stamp.ccsds_days, 1)
         self.assertEqual(stamp.ms_of_day, 5)
+
+    def test_dt_is_utc(self):
+        empty_stamp = CdsShortTimestamp(0, 0)
+        dt = empty_stamp.as_date_time()
+        self.assertEqual(dt.tzinfo, datetime.timezone.utc)
+        self.assertEqual(dt.tzinfo.utcoffset(dt), datetime.timedelta(0))
+
+    def test_compare_from_now_against_manually_created(self):
+        stamp = CdsShortTimestamp.from_now()
+        ccsds_days = stamp.ccsds_days
+        ms_of_day = stamp.ms_of_day
+        new_stamp = CdsShortTimestamp(ccsds_days, ms_of_day)
+        self.assertEqual(new_stamp, stamp)
+        self.assertEqual(new_stamp.as_unix_seconds(), new_stamp.as_unix_seconds())